+++ src/menu.rs
@@ -4,7 +4,7 @@ use leafwing_input_manager::{
     Actionlike,
 };
 
-use crate::state::AppState;
+use crate::state::{is_in_menu_state, AppState};
 
 use crate::effects::DrawBlinkTimer;
 
@@ -14,23 +14,25 @@ pub struct StartMenuScreen {}
 #[derive(Actionlike, PartialEq, Eq, Clone, Copy, Hash, Debug, Reflect)]
 pub enum MenuAction {
     Start,
+    Credits,
 }
 
 pub struct MenuPlugin;
 impl Plugin for MenuPlugin {
     fn build(&self, app: &mut App) {
-        app.add_systems(Startup, setup)
-            .add_systems(
-                Update,
-                menu_input_system.run_if(in_state(AppState::StartMenu)),
-            )
+        app.add_systems(OnEnter(AppState::StartMenu), setup)
+            .add_systems(Update, menu_input_system.run_if(is_in_menu_state))
             .add_systems(OnExit(AppState::StartMenu), despawn);
     }
 }
 
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
-    let mut input_map = InputMap::<MenuAction>::new([(KeyCode::Return, MenuAction::Start)]);
+    let mut input_map = InputMap::<MenuAction>::new([
+        (KeyCode::Return, MenuAction::Start),
+        (KeyCode::C, MenuAction::Credits),
+    ]);
     input_map.insert(GamepadButtonType::Start, MenuAction::Start);
+    input_map.insert(GamepadButtonType::North, MenuAction::Credits);
 
     commands.insert_resource(input_map.build());
     commands.insert_resource(ActionState::<MenuAction>::default());
@@ -80,6 +82,47 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
             ));
         });
 
+    commands
+        .spawn((
+            NodeBundle {
+                style: Style {
+                    position_type: PositionType::Absolute,
+                    width: Val::Percent(100.0),
+                    height: Val::Percent(100.0),
+                    align_items: AlignItems::FlexEnd,
+                    justify_content: JustifyContent::FlexEnd,
+                    flex_direction: FlexDirection::Row,
+                    ..default()
+                },
+                ..default()
+            },
+            StartMenuScreen {},
+        ))
+        .with_children(|parent| {
+            parent.spawn((TextBundle {
+                style: Style {
+                    justify_content: JustifyContent::FlexEnd,
+                    margin: UiRect {
+                        left: Val::Px(10.0),
+                        right: Val::Px(10.0),
+                        top: Val::Px(10.0),
+                        bottom: Val::Px(10.0),
+                    },
+                    ..default()
+                },
+                text: Text::from_section(
+                    "Press 'C' for Credits",
+                    TextStyle {
+                        font: asset_server.load("fonts/kenvector_future.ttf"),
+                        font_size: 25.0,
+                        color: Color::rgb_u8(0xAA, 0xAA, 0x33),
+                        ..default()
+                    },
+                ),
+                ..default()
+            },));
+        });
+
     commands.spawn((
         AudioBundle {
             source: asset_server.load("sound/Beat Mekanik - Lightspeed.ogg"),
@@ -99,10 +142,22 @@ fn despawn(mut commands: Commands, query: Query<Entity, With<StartMenuScreen>>) 
 }
 
 fn menu_input_system(
+    state: Res<State<AppState>>,
     mut next_state: ResMut<NextState<AppState>>,
     inputs: Res<ActionState<MenuAction>>,
 ) {
     if inputs.just_pressed(MenuAction::Start) {
         next_state.set(AppState::GameCreate);
+    }
+    if inputs.just_released(MenuAction::Credits) {
+        match state.get() {
+            AppState::StartMenu => {
+                next_state.set(AppState::Credits);
+            }
+            AppState::Credits => {
+                next_state.set(AppState::StartMenu);
+            }
+            _ => {}
+        }
     }
 }