+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Start menu
 - Start menu music: Lightspeed by Beat Mekanik (Free Music Archive, CC BY)
 - Ambient music: Space Dust by Kirk Osamayo (Free Music Archive, CC BY)
+- In-game credits accessible from the start menu
 - Download links and controls in README
 
 ### Changed